switch to pynput for cross-platform
Changing out evdev for pynput for cross-platform support. This introduces some annoyances for Linux users, namely that by default, xorg is relied upon for sending keys, meaning that it'll only work under an X environment or in X applications running on Wayland under Xwayland. Most games tend to run as an X application anyways, so it shouldn't be an issue for most people. If you play osu! you may need to make sure the SDL_VIDEODRIVER envvar is not set to Wayland.

You can potentially get around this for Wayland applications by changing the backend to uinput with the envvar $PYNPUT_BACKEND_KEYBOARD and running the script as root, but it feels like overkill to me (though I suppose it's better than just being allowed to do all this in X without a care in the world?)

https://pynput.readthedocs.io/en/latest/index.html?highlight=backend